Moved pkg/core/containercreator to pkg/containercreator
