+++ pkg/containercreator/docker/docker.go
@@ -1,8 +1,8 @@
 package docker
 
 import (
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator"
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator/utils"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator/utils"
 	kubecore "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 )
