+++ pkg/containercreator/docker/variables.go
@@ -5,9 +5,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator"
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator/kaniko"
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator/utils"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator/kaniko"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator/utils"
 	"github.com/iver-wharf/wharf-core/pkg/logger"
 )
