+++ pkg/containercreator/docker/variabletype.go
@@ -4,8 +4,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator"
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator/kaniko"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator/kaniko"
 )
 
 type VariableType int
