+++ pkg/containercreator/helm/helm.go
@@ -1,7 +1,7 @@
 package helm
 
 import (
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator/utils"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator/utils"
 )
 
 const (
