+++ pkg/containercreator/kaniko/kaniko.go
@@ -1,7 +1,7 @@
 package kaniko
 
 import (
-	"github.com/iver-wharf/wharf-cmd/pkg/core/containercreator/utils"
+	"github.com/iver-wharf/wharf-cmd/pkg/containercreator/utils"
 )
 
 const (
