+++ pkg/containercreator/utils/scriptworker.go
