+++ pkg/containercreator/utils/scriptworker_test.go
