+++ pkg/containercreator/utils/utils.go
